GCP_Webhooks/google_translate/translate/requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
cloudevents==1.10.1
deprecation==2.1.0
flask==3.0.0; python_version >= '3.8'
functions-framework==3.5.0
google-api-core[grpc]==2.15.0; python_version >= '3.7'
google-auth==2.25.1; python_version >= '3.7'
google-cloud==0.34.0
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-dialogflow-cx==1.30.0
google-cloud-translate==3.13.0
googleapis-common-protos==1.62.0; python_version >= '3.7'
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0; platform_system != 'Windows'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.1; python_version >= '3.8'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
urllib3==2.1.0; python_version >= '3.8'
watchdog==3.0.0; python_version >= '3.7'
werkzeug==3.0.1; python_version >= '3.8'
